Parallel version.
The point here is, so far as possible, to be able to run the tests in
parallel using https://github.com/simonjwright/parallelize. The tests
can still be run serially if required.

parallelize requires a separate script for each job to be run in
parallel; here, we create a separate script for each test. The tests
for each chapter are run in parallel.

Because of gnatchop's handling of files containing just configuration
pragmas (they're generated in gnat.adc in the current working
directory; see GCC PR 113877), each test is run in its own
directory. This means that we have one script and one directory per
test! (the directories are deleted after the test completes, but the
script remains).

  * gnatscrp.a (Extra_Compiler_Options): added -g. support/ is one level
      up.
    (Unit_Script_File): used to generate individual test scripts (the
      original Script_File calls all the tests serially).
    (Generate_Test_Compilation_Script): remove all the redirection and
      appending of standard in/out to the result file; parallelize does
      this.
      Make adjustments to the test script to create the test subdirectory
      and run it there.
    (reported version): 2.0.
    (Create_Unit_Script_File): new.
  * setup-working-directory.sh: after gnatchopping all the files from
      the ACATS support/ directory, edit report.adb so that the file
      specs created by Report.Legal_File_Name are prefixed by
      '../'. This is so that files intended to be created in one test
      and checked in a subsequent one are preserved after the first
      test's directory is deleted. This causes issues with CXAG001,
      CXAG002.