Retain "legal file names" between tests.
Some tests involve creating files, and the names for those files are
generated by Report.Legal_File_Name.

For most tests, those files can live in the test's working
directory. There are a few where the file needs to be retained
between tests (test A creates the file, test B checks to see
whether a previously-created file can be opened).

We do this by arranging for the files to be created in the parent
directory of the test's working directory.

  * report-legal_file_name.adb: new. This alternative version of the
      equivalent in ACATS/support prepends "../" to the generated name
      for the tests that require this treatment (CE2108E, CE2108G).
  * setup-working-directory.sh: after our copy of the ACATS version of
      report-legal_file_name.adb has been generated, copies our version
      over it.
  * .gitignore: ignore all subdirectories.